chore!: migrate to vite
