+++ .nvmrc
@@ -1,1 +1,1 @@
-v18
+lts/*
