+++ src/components/Todo.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef, useState } from "react";
+import { useEffect, useRef, useState } from "react";
 
 function usePrevious(value) {
   const ref = useRef(null);
