+++ src/index.css
@@ -4,7 +4,7 @@
 *::after {
   box-sizing: border-box;
 }
-*:focus {
+*:focus-visible {
   outline: 3px dashed #228bec;
   outline-offset: 0;
 }
@@ -20,18 +20,18 @@ ul {
   padding: 0;
 }
 button {
-  border: none;
-  margin: 0;
-  padding: 0;
-  width: auto;
-  overflow: visible;
+  -moz-osx-font-smoothing: inherit;
+  -webkit-font-smoothing: inherit;
+  appearance: none;
   background: transparent;
+  border: none;
   color: inherit;
   font: inherit;
   line-height: normal;
-  -webkit-font-smoothing: inherit;
-  -moz-osx-font-smoothing: inherit;
-  -webkit-appearance: none;
+  margin: 0;
+  overflow: visible;
+  padding: 0;
+  width: auto;
 }
 button::-moz-focus-inner {
   border: 0;
@@ -54,12 +54,14 @@ input[type="text"] {
   border-radius: 0;
 }
 body {
-  width: 100%;
-  max-width: 68rem;
-  margin: 0 auto;
-  font: 1.6rem/1.25 Arial, sans-serif;
   background-color: #f5f5f5;
   color: #4d4d4d;
+  font:
+    1.6rem/1.25 Arial,
+    sans-serif;
+  margin: 0 auto;
+  max-width: 68rem;
+  width: 100%;
 }
 @media screen and (min-width: 620px) {
   body {
@@ -74,30 +76,30 @@ body {
   margin-top: 0.4rem;
 }
 .btn {
-  padding: 0.8rem 1rem 0.7rem;
   border: 0.2rem solid #4d4d4d;
   cursor: pointer;
+  padding: 0.8rem 1rem 0.7rem;
   text-transform: capitalize;
 }
 .btn.toggle-btn {
-  border-width: 1px;
   border-color: #d3d3d3;
+  border-width: 1px;
 }
 .btn.toggle-btn[aria-pressed="true"] {
-  text-decoration: underline;
   border-color: #4d4d4d;
+  text-decoration: underline;
 }
 .btn__danger {
-  color: #fff;
   background-color: #ca3c3c;
   border-color: #bd2130;
+  color: #fff;
 }
 .btn__filter {
   border-color: lightgrey;
 }
 .btn__primary {
-  color: #fff;
   background-color: #000;
+  color: #fff;
 }
 .btn-group {
   display: flex;
@@ -110,22 +112,22 @@ body {
   margin-left: 0.8rem;
 }
 .label-wrapper {
-  margin: 0;
   flex: 0 0 100%;
+  margin: 0;
   text-align: center;
 }
 .visually-hidden {
-  position: absolute !important;
-  height: 1px;
-  width: 1px;
-  overflow: hidden;
   clip: rect(1px 1px 1px 1px);
   clip: rect(1px, 1px, 1px, 1px);
+  height: 1px;
+  overflow: hidden;
+  position: absolute !important;
   white-space: nowrap;
+  width: 1px;
 }
 [class*="stack"] > * {
-  margin-top: 0;
   margin-bottom: 0;
+  margin-top: 0;
 }
 .stack-small > * + * {
   margin-top: 1.25rem;
@@ -147,10 +149,12 @@ body {
 /* END GLOBAL STYLES */
 .todoapp {
   background: #fff;
+  box-shadow:
+    0 2px 4px 0 rgba(0, 0, 0, 0.2),
+    0 2.5rem 5rem 0 rgba(0, 0, 0, 0.1);
   margin: 2rem 0 4rem 0;
   padding: 1rem;
   position: relative;
-  box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.2), 0 2.5rem 5rem 0 rgba(0, 0, 0, 0.1);
 }
 @media screen and (min-width: 550px) {
   .todoapp {
@@ -158,39 +162,39 @@ body {
   }
 }
 .todoapp > * {
-  max-width: 50rem;
   margin-left: auto;
   margin-right: auto;
+  max-width: 50rem;
 }
 .todoapp > form {
   max-width: 100%;
 }
 .todoapp > h1 {
   display: block;
-  max-width: 100%;
-  text-align: center;
   margin: 0;
   margin-bottom: 1rem;
+  max-width: 100%;
+  text-align: center;
 }
 .label__lg {
   line-height: 1.01567;
   font-weight: 300;
-  padding: 0.8rem;
   margin-bottom: 1rem;
+  padding: 0.8rem;
   text-align: center;
 }
 .input__lg {
-  padding: 2rem;
   border: 2px solid #000;
+  padding: 2rem;
 }
-.input__lg:focus {
+.input__lg:focus-visible {
   border-color: #4d4d4d;
   box-shadow: inset 0 0 0 2px;
 }
 [class*="__lg"] {
   display: inline-block;
-  width: 100%;
   font-size: 1.9rem;
+  width: 100%;
 }
 [class*="__lg"]:not(:last-child) {
   margin-bottom: 1rem;
@@ -210,78 +214,78 @@ body {
   flex: 0 0 100%;
 }
 .todo-text {
-  width: 100%;
+  border: 2px solid #565656;
   min-height: 4.4rem;
   padding: 0.4rem 0.8rem;
-  border: 2px solid #565656;
+  width: 100%;
 }
-.todo-text:focus {
+.todo-text:focus-visible {
   box-shadow: inset 0 0 0 2px;
 }
 /* CHECKBOX STYLES */
 .c-cb {
-  box-sizing: border-box;
-  font-family: Arial, sans-serif;
   -webkit-font-smoothing: antialiased;
-  font-weight: 400;
-  font-size: 1.6rem;
-  line-height: 1.25;
+  box-sizing: border-box;
+  clear: left;
   display: block;
-  position: relative;
+  font-family: Arial, sans-serif;
+  font-size: 1.6rem;
+  font-weight: 400;
+  line-height: 1.25;
   min-height: 44px;
   padding-left: 40px;
-  clear: left;
+  position: relative;
 }
 .c-cb > label::before,
 .c-cb > input[type="checkbox"] {
   box-sizing: border-box;
-  top: -2px;
-  left: -2px;
-  width: 44px;
   height: 44px;
+  left: -2px;
+  top: -2px;
+  width: 44px;
 }
 .c-cb > input[type="checkbox"] {
   -webkit-font-smoothing: antialiased;
   cursor: pointer;
-  position: absolute;
-  z-index: 1;
   margin: 0;
   opacity: 0;
+  position: absolute;
+  z-index: 1;
 }
 .c-cb > label {
-  font-size: inherit;
-  font-family: inherit;
-  line-height: inherit;
+  cursor: pointer;
   display: inline-block;
+  font-family: inherit;
+  font-size: inherit;
+  line-height: inherit;
   margin-bottom: 0;
   padding: 8px 15px 5px;
-  cursor: pointer;
   touch-action: manipulation;
 }
 .c-cb > label::before {
+  background: transparent;
+  border: 2px solid currentcolor;
   content: "";
   position: absolute;
-  border: 2px solid currentColor;
-  background: transparent;
 }
-.c-cb > input[type="checkbox"]:focus + label::before {
+.c-cb > input[type="checkbox"]:focus-visible + label::before {
   border-width: 4px;
   outline: 3px dashed #228bec;
 }
 .c-cb > label::after {
-  box-sizing: content-box;
-  content: "";
-  position: absolute;
-  top: 11px;
-  left: 9px;
-  width: 18px;
-  height: 7px;
-  transform: rotate(-45deg);
+  background: transparent;
   border: solid;
   border-width: 0 0 5px 5px;
   border-top-color: transparent;
+  box-sizing: content-box;
+  content: "";
+  height: 7px;
+  left: 9px;
   opacity: 0;
-  background: transparent;
+  position: absolute;
+  top: 11px;
+  transform: rotate(-45deg);
+  width: 18px;
 }
 .c-cb > input[type="checkbox"]:checked + label::after {
   opacity: 1;
